Replaced iterator and generating function constructors with vectors
After about a year of this design pattern, it was found that
the generating functions were rarely used and that the iterators
almost always were vector iterators. In addition, using vectors
will allow for move semantics to potentially save memory in some
cases. While the removal of generating functions no longer allows for
O(1) construction in some cases, it has come up very rarely. These
changes result in shorter code in most cases, which is an added bonus.